move getSavedFilePaths to service
