+++ src/services/upload.ts
@@ -1,7 +1,7 @@
 import { setToUploadFiles } from '../api/upload';
-import { getValidPaths } from '../services/fs';
 import { uploadStatusStore } from '../stores/upload.store';
 import { FILE_PATH_TYPE, FILE_PATH_KEYS } from '../types';
+import { getValidPaths } from './fs';
 
 export const getSavedFilePaths = (type: FILE_PATH_TYPE) => {
     const paths =
